chore(web): rename IP comment edit files to match existing convention
fix(web): exit early from post review controller if there are errors

fix(web): enable success banner after changing address on IP comment

fix(web): further copy updates

test(web): update test snapshots
