+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/redact/__tests__/redact.test.js
@@ -40,7 +40,7 @@ describe('redact', () => {
 		const { innerHTML } = parseHtml(response.text);
 		expect(innerHTML).toMatchSnapshot();
 
-		expect(innerHTML).toContain('Redact comment from Lee Thornton');
+		expect(innerHTML).toContain('Check details and redact comment');
 		expect(innerHTML).toContain('Awaiting review comment 47');
 	});
