+++ appeals/web/src/server/appeals/appeal-details/interested-party-comments/view-and-review/view-and-review.controller.js
@@ -61,7 +61,7 @@ export const postReviewInterestedPartyComment = async (request, response, next) 
 		}
 
 		if (errors) {
-			renderViewInterestedPartyComment(request, response, next);
+			return renderViewInterestedPartyComment(request, response, next);
 		}
 
 		if (status === COMMENT_STATUS.VALID_REQUIRES_REDACTION) {
