+++ appeals/web/src/server/lib/mappers/components/page-components/notification-banners.mapper.js
@@ -208,12 +208,12 @@ export const notificationBannerDefinitions = {
 	},
 	interestedPartyCommentsAddressAddedSuccess: {
 		type: 'success',
-		pages: ['viewIpComment'],
+		pages: ['viewIpComment', 'reviewIpComment'],
 		text: 'Interested party address added'
 	},
 	interestedPartyCommentsAddressUpdatedSuccess: {
 		type: 'success',
-		pages: ['viewIpComment'],
+		pages: ['viewIpComment', 'reviewIpComment'],
 		text: 'Interested party address changed'
 	},
 	interestedPartyCommentsDocumentAddedSuccess: {
